制定本次升级的目标，增加freemarker生成代码的工具类，将Double类型替换成BigDecimal类型

+++ src/main/java/com/linshang/code/common/PropertyNameTools.java
@@ -18,9 +18,9 @@ public class PropertyNameTools {
 		typeMap.put("INT", "Long");
 		typeMap.put("BIT", "Boolean");
 		
-		typeMap.put("FLOAT", "Double");
-		typeMap.put("DOUBLE", "Double");
-		typeMap.put("DECIMAL", "Double");
+		typeMap.put("FLOAT", "BigDecimal");
+		typeMap.put("DOUBLE", "BigDecimal");
+		typeMap.put("DECIMAL", "BigDecimal");
 		
 		typeMap.put("CHAR", "String");
 		typeMap.put("VARCHAR", "String");
